% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofarR_functions.R
\name{cycling_summary()}
\alias{cycling_summary()}
\alias{cycling_summary}
\title{Cycling Summary}
\usage{
cycling_summary(data = workout_data_interface, type = "All", min_time = 11, min_output = 50)
}
\arguments{
\item{data}{- A formatted dataframe read in using setup_peloton_data. Defaults to a variable, "workout_data_interface", that is the output from that function.}

\item{type}{Which type of}

\item{min_time}{The minimum amount of time to be summarized, used to exclude short rides, either ones that were aborted or warmups/cooldowns. Defaults to 11 minutes to exclude all rides 10  minutes or shorter}

\item{min_output}{The minimum output to be included in the summary. Defaults to 50 KJ}
}
\value{

}
\description{
Cycling Summary
}
\examples{

cycling_summary()
cycling_summary(type = "Climb", min_time = 11, min_ouput = 200)

}
