% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofarR_functions.R
\name{decoupling()}
\alias{decoupling()}
\alias{decoupling}
\title{Decoupling Analysis}
\usage{
decoupling(data = NULL)
}
\arguments{
\item{data}{Data pulled from the API, should be stored differently than the data pulled from the interface.}
}
\description{
This function takes watt and heartrate data from the API, subsets it to only look at the workouts that you specify, and determines your decoupling by workout, as well as trends by workout length and by time.
}
