% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofarR_functions.R
\name{setup_peloton_data()}
\alias{setup_peloton_data()}
\alias{setup_peloton_data}
\title{Setup Peloton data

Set up peloton data, which you can retrieve directly from your profile once you've logged in to onepeloton.com. Assumes that your workout data has not been renamed and is in your working directory.}
\usage{
setup_peloton_data(display_name)
}
\arguments{
\item{display_name}{Your display name on the leaderboard. If you have already set it as a string in a variable named \code{leaderboard_name}, it will use that by default}
}
\value{
Returns a dataframe, workout_data_interface. Please save it in your global environment as under the same name, workout_data_interface, since that is the default for all other functions in this package.
}
\description{
Setup Peloton data

Set up peloton data, which you can retrieve directly from your profile once you've logged in to onepeloton.com. Assumes that your workout data has not been renamed and is in your working directory.
}
