% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofarR_functions.R
\name{setup_peloton_data}
\alias{setup_peloton_data}
\title{Setup Peloton data

Set up peloton data, which you can retrieve directly from your profile once you've logged in to onepeloton.com}
\usage{
setup_peloton_data(display_name)
}
\arguments{
\item{display_name}{Your display name on the leaderboard. If you have already set it as a string in a variable named \code{leaderboard_name}, it will use that by default}
}
\value{
Returns a dataframe, workout_data_interface
}
\description{
Setup Peloton data

Set up peloton data, which you can retrieve directly from your profile once you've logged in to onepeloton.com
}
