% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gofarR_functions.R
\name{top_instructor()}
\alias{top_instructor()}
\alias{top_instructor}
\title{View top instructor}
\usage{
top_instructor(w_data = workout_data_interface, discipline = "All")
}
\arguments{
\item{w_data}{- Workout data read in from setup_peloton_data}

\item{discipline}{The class discipline to be summarized. Defaults to "All"}
}
\description{
View top instructor
}
\examples{
top_instructor(workout_data_interface)
top_instructor(discipline = "Yoga")

}
